changes are done for all items like comment,wo

+++ Utilities/ImageUploadInFile.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Http;
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using Utilities.Interface;
@@ -8,11 +9,13 @@ namespace Utilities
 {
     public class ImageUploadInFile : IImageUploadInFile
     {
+        private List<string> allowedTypes=new List<string> {".jpg",".jpeg",".png",".pdf",".xls",".xlx",".doc",".docx" };
+        
         public bool Delete(string path)
         {
             try
             {
-                if (File.Exists(path))
+                if (!string.IsNullOrWhiteSpace(path) && File.Exists(path))
                 {
                     File.Delete(path);
                     return true;
@@ -29,10 +32,14 @@ namespace Utilities
         public async Task<string> UploadAsync(IFormFile file)
         {
             var id = Guid.NewGuid().ToString().Replace("-", "");
-            if (file.Length > 0)
+            if (file!=null && file.Length > 0)
             {
+
                 try
                 {
+                    var type=System.IO.Path.GetExtension(file.FileName);
+                    if (!allowedTypes.Contains(type))
+                        return null;
                     if (!Directory.Exists("ImageFileStore"))
                     {
                         Directory.CreateDirectory("ImageFileStore");
